+++ app/src/main/java/uk/org/openseizuredetector/OsdUtil.java
@@ -96,6 +96,7 @@ public class OsdUtil implements ActivityCompat.OnRequestPermissionsResultCallbac
     private final String[] SMS_PERMISSIONS = {
             Manifest.permission.SEND_SMS,
             Manifest.permission.ACCESS_FINE_LOCATION,
+            Manifest.permission.READ_PHONE_STATE,
     };