+++ app/src/main/java/uk/org/openseizuredetector/SdServer.java
@@ -629,7 +629,7 @@ public class SdServer extends Service implements SdDataReceiver {
         }
 
         // Fault
-        if ((sdData.alarmState) == 4 || (sdData.alarmState == 7)) {
+        if ((sdData.alarmState) == 4 || (sdData.alarmState == 7) || (sdData.mHRFaultStanding)) {
             sdData.alarmPhrase = "FAULT";
             writeAlarmToSD();
             faultWarningBeep();