+++ app/src/main/java/uk/org/openseizuredetector/StartupActivity.java
@@ -340,7 +340,7 @@ public class StartupActivity extends Activity {
             // Is Pebble Watch App Running?
             tv = (TextView) findViewById(R.id.textItem4);
             pb = (ProgressBar) findViewById(R.id.progressBar4);
-            if (mConnection.pebbleAppRunning()) {
+            if (mConnection.watchAppRunning()) {
                 tv.setText("Watch App Running OK");
                 tv.setBackgroundColor(okColour);
                 tv.setTextColor(okTextColour);
@@ -506,6 +506,7 @@ public class StartupActivity extends Activity {
                             + "\n  V3.1.5 - Added repeat alarm beeps during SMS delay to alert user.."
                             + "\n  V3.1.6 - Made Cancel Audible button inhibit sending SMS alarms as well as audible beeps"
                             + "\n  V3.1.8 - Added READ_PHONE_STATE permission, which seems to be needed for some phones"
+                            + "\n  V3.1.9 - Fixed issue with Garmin Seizure Detector not producing warnings.  Added faut pips for missing heart rate data if heart rate alarm active"
                 );
             // This makes the links display as links, but they do not respond to clicks for some reason...
             Linkify.addLinks(s, Linkify.ALL);
@@ -540,6 +541,7 @@ public class StartupActivity extends Activity {
                             + "\n  V3.1.5 - Added repeat alarm beeps during SMS delay to alert user.."
                             + "\n  V3.1.6 - Made Cancel Audible button inhibit sending SMS alarms as well as audible beeps"
                             + "\n  V3.1.8 - Added READ_PHONE_STATE permission, which seems to be needed for some phones"
+                            + "\n  V3.1.9 - Fixed issue with Garmin Seizure Detector not producing warnings.  Added faut pips for missing heart rate data if heart rate alarm active"
                             + "\n "
             );
             // This makes the links display as links, but they do not respond to clicks for some reason...